requirements.txt
pypdf
sentence-transformers
faiss-cpu
fastapi
uvicorn
streamlit
numpy
python-multipart
